zougsdig

+++ my_average.c
@@ -1,5 +1,5 @@
 float my_average(int *grades, int count) {
-    long long int total = 0;
+    int total = 0;
     double average;
 
     for (int i = 0; i<count; i++) {
